chore(dev): Add .tool-versions for Ruby version management (#54)
* chore(dev): Add .tool-versions for Ruby version management

This commit introduces a new file called .tool-versions. It's used to specify the Ruby version (2.7.2) needed for the project, essential for consistent development environment setup across different machines.

* chore: Update README to reference Ruby and Rails versions

Instead of hardcoding the Ruby and Rails versions in the README, they are now referred to in the .tool-versions and Gemfile respectively. This ensures that the README stays updated with the actual versions used in the project.